Added EditInterventionsPage and CreateNewClient page.
- Added EditInterventionPage - when Edit is pressed on Interventions or
ProposedInterventions, the selected index of the row in the table is
sent via Session[], and is displayed in the Edit Intervention Page. (Not
working - need selectable table rows - sends 0 all the time).
The buttons for editing display only appear if the user has correct
permissions - and the quality information only appears if it exists.
(Not fully implemented, due to no ability to get current user info.)

- Implemented Create New Client page - when the user clicks the Create
New... button, they can fill out name, location and choose a District.
It gets added to the Clients list.

- Implemented TreeNode module for Clients page - it now populates
dynamically from the Clients list.

- Re-factored code in Interventions and a few other places.